pyproject.toml
# pyproject.toml
[project]
name = "youtube-chatbot"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langsmith>=0.1.0",
    "nemoguardrails>=0.10.0",
    "redis>=5.0.0",
    "rq>=1.16.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "httpx>=0.27.0",
    "youtube-transcript-api>=0.6.0",
    "google-api-python-client>=2.140.0",
    "pydantic>=2.8.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
